NAT-1: load project and draw images

+++ nautilus-annotation-tool/src/main/java/com/nautilus/nat/model/BoundingBox.java
@@ -3,10 +3,10 @@ package com.nautilus.nat.model;
 public class BoundingBox {
   private int classIndex;
   private String className;
-  private float x;
-  private float y;
-  private float width;
-  private float height;
+  private double x;
+  private double y;
+  private double width;
+  private double height;
 
   public interface BoundingBoxChangeListener {
     void onPropertyChanged(BoundingBox bbox, Object notifier);
@@ -28,35 +28,35 @@ public class BoundingBox {
     this.className = className;
   }
 
-  public float getX() {
+  public double getX() {
     return x;
   }
 
-  public void setX(float x) {
+  public void setX(double x) {
     this.x = x;
   }
 
-  public float getY() {
+  public double getY() {
     return y;
   }
 
-  public void setY(float y) {
+  public void setY(double y) {
     this.y = y;
   }
 
-  public float getWidth() {
+  public double getWidth() {
     return width;
   }
 
-  public void setWidth(float width) {
+  public void setWidth(double width) {
     this.width = width;
   }
 
-  public float getHeight() {
+  public double getHeight() {
     return height;
   }
 
-  public void setHeight(float height) {
+  public void setHeight(double height) {
     this.height = height;
   }
 }
